Move currentUser and generalSettings queries to shared/data
