--- assets/prototype/domain/eventEditor/data/queries/generalSettings/index.ts
@@ -1,11 +1,0 @@
-import gql from 'graphql-tag';
-
-export const GET_GENERAL_SETTINGS: any = gql`
-	query GET_GENERAL_SETTINGS {
-		generalSettings {
-			dateFormat
-			timeFormat
-			timezone
-		}
-	}
-`;
