+++ assets/prototype/domain/shared/data/queries/generalSettings/useGeneralSettings.ts
@@ -2,7 +2,7 @@ import pathOr from 'ramda/src/pathOr';
 import { useApolloClient } from '@apollo/react-hooks';
 
 import { GET_GENERAL_SETTINGS } from './';
-import { ReadQueryOptions } from '../types';
+import { ReadQueryOptions } from '../../../../eventEditor/data/queries/types';
 import { GeneralSettings, GeneralSettingsData } from '../../../../../application/valueObjects/config/types';
 /**
  * A custom react hook for retrieving GeneralSettings
